Add supplicant device to handle supplicant cases
The code without a special supplicant device requires introduciton of
new IOCTLs which cause code repetition and increases complexity. A new device
/dev/tee-supp0 was created which modified ioctl_register and alloc to see if we
are a supplicant or not, so no new ioctl was introduced.

The shm_list is split to private and supplicant. The supplicant list contains
all shm allocations that happened from supplicant context. That way a process
the closes the /dev/tee0, will cleanup all the local shm but will not touch the
ones that happened by the supplicant. The memory held by the supplicant will be
removed either with an rpc request, or when the process closes /dev/tee-supp0